+++ libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: amazurie <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/03 17:29:42 by amazurie          #+#    #+#             */
-/*   Updated: 2017/01/30 12:09:13 by amazurie         ###   ########.fr       */
+/*   Updated: 2017/03/15 12:32:10 by amazurie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -89,7 +89,7 @@ int				ft_lstlen(t_list **alst);
 char			*ft_strtrimc(const char *s, char c);
 char			*ft_strndup(const char *s1, size_t n);
 char			*ft_strnjoin(char const *s1, char const *s2, size_t n);
-int				ft_strlen_chr(char *s, char c);
+size_t			ft_strlen_chr(char *s, char c);
 void			ft_putnstr(char const *str, size_t str_len);
 void			ft_putnchar(const char c, size_t n);
 size_t			ft_lennb(size_t c);
